requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
python-dotenv==1.0.1
werkzeug==3.0.4
bleach==6.0.0
python-dateutil==2.8.2
openpyxl==3.1.2
python-telegram-bot==20.8
requests==2.31.0
apscheduler==3.10.4
cryptography==42.0.8